* DiscImageChef.Filesystems/UCSDPascal/Dir.cs: 	  Typo.
	* DiscImageChef.Helpers/DateHandlers.cs:
	  Added CP/M timestamp converter.

	* DiscImageChef.Partitions/Acorn.cs:
	  Corrected handling of negative values.

	* DiscImageChef/Commands/ExtractFiles.cs:
	  Corrected behaviour when volume name is missing, null or
	  empty.

	* DiscImageChef.DiscImages/ImagePlugin.cs:
	  Added floppy address mark sector tag.

+++ DiscImageChef.Partitions/Acorn.cs
@@ -86,7 +86,7 @@ namespace DiscImageChef.PartPlugins
             int secCyl = bootBlock.discRecords.spt * heads;
             int mapSector = bootBlock.startCylinder * secCyl;
 
-            if(mapSector >= (int)imagePlugin.GetSectors())
+            if((ulong)mapSector >= imagePlugin.GetSectors())
                 return false;
 
             byte[] map = imagePlugin.ReadSector((ulong)mapSector);
